+++ src/django_mysql/models/lookups.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
+from collections.abc import Iterable
 from typing import Any
 from typing import Callable
-from typing import Iterable
 
 from django.db.backends.base.base import BaseDatabaseWrapper
 from django.db.models import CharField