+++ tests/requirements/compile.py
@@ -27,46 +27,6 @@ if __name__ == "__main__":
         [
             *common_args,
             "--python",
-            "3.8",
-            "--output-file",
-            "py38-django32.txt",
-        ],
-        input=b"Django>=3.2a1,<3.3\nmysqlclient>=1.4.0\n",
-    )
-    run(
-        [
-            *common_args,
-            "--python",
-            "3.8",
-            "--output-file",
-            "py38-django40.txt",
-        ],
-        input=b"Django>=4.0a1,<4.1\nmysqlclient>=1.4.0\n",
-    )
-    run(
-        [
-            *common_args,
-            "--python",
-            "3.8",
-            "--output-file",
-            "py38-django41.txt",
-        ],
-        input=b"Django>=4.1a1,<4.2\nmysqlclient>=1.4.0\n",
-    )
-    run(
-        [
-            *common_args,
-            "--python",
-            "3.8",
-            "--output-file",
-            "py38-django42.txt",
-        ],
-        input=b"Django>=4.2a1,<5.0\nmysqlclient>=1.4.3\n",
-    )
-    run(
-        [
-            *common_args,
-            "--python",
             "3.9",
             "--output-file",
             "py39-django32.txt",